Refactor storage of mountebank methods

+++ mongoDBImpostersRepository.js
@@ -1,16 +1,53 @@
 'use strict';
 
-const { MongoClient, Code } = require('mongodb');
+const { MongoClient } = require('mongodb');
 const errors = require('./utils/errors');
 
 /* TODO:
  * Error handling
- * Address async/sync issues with stubsFor and stopAllSync
 */
 function create (config, logger) {
   const mongoCfg = getMongoConfig(config, logger),
     client = new MongoClient(mongoCfg.uri);
   client.connect();
+
+  const imposterFns = {};
+
+  /**
+   * Saves a reference to the imposter so that the functions
+   * (which can't be persisted) can be rehydrated to a loaded imposter.
+   * This means that any data in the function closures will be held in
+   * memory.
+   * @memberOf module:mongoDBImpostersRepository#
+   * @param {Object} imposter - the imposter
+   */
+  function addReference (imposter) {
+    const id = String(imposter.port);
+    imposterFns[id] = {};
+    Object.keys(imposter).forEach(key => {
+        if (typeof imposter[key] === 'function') {
+            imposterFns[id][key] = imposter[key];
+        }
+    });
+  }
+
+  /**
+   * Functions saved in addReference method for each imposter
+   * is added back to the imposter
+   * @memberOf module:mongoDBImpostersRepository#
+   * @param {Object} imposter - the imposter
+   * @returns {Object} - the promise
+   */
+  function rehydrate (imposter) {
+    const id = String(imposter.port);
+    if (!imposterFns[id]) {
+      return imposter;
+    }
+    Object.keys(imposterFns[id]).forEach(key => {
+        imposter[key] = imposterFns[id][key];
+    });
+    return imposter;
+  }
 
   /**
    * Adds a new imposter
@@ -19,13 +56,15 @@ function create (config, logger) {
    * @returns {Object} - the promise
    */
   async function add (imposter) {
-    if (!imposter.stubs) {
-      imposter.stubs = [];
+    if (!imposter.creationRequest.stubs) {
+      imposter.creationRequest.stubs = [];
     }
 
     const doc = {};
-    doc[imposter.port] = imposter;
-    await client.db(mongoCfg.db).collection('imposters').insertOne(doc, { serializeFunctions: true });
+    doc[imposter.port] = imposter.creationRequest;
+    await client.db(mongoCfg.db).collection('imposters').insertOne(doc);
+
+    addReference(imposter);
     return imposter;
   }
 
@@ -34,7 +73,7 @@ function create (config, logger) {
       options = {};
     doc[imposter.port] = imposter;
     options[imposter.port] = { $exists: true };
-    await client.db(mongoCfg.db).collection('imposters').replaceOne(options, doc, { serializeFunctions: true });
+    await client.db(mongoCfg.db).collection('imposters').replaceOne(options, doc);
   }
 
   /**
@@ -48,16 +87,12 @@ function create (config, logger) {
     const database = client.db(mongoCfg.db),
       options = {};
     options[id] = { $exists: true };
-    result = await database.collection('imposters').findOne(options, { serializeFunctions: true });
+    result = await database.collection('imposters').findOne(options);
 
     if (result) {
-      Object.keys(result[String(id)]).forEach(k => {
-        if (result[String(id)][k] instanceof Code) {
-          // dear god, kill me T_T (╯°□°)╯︵ ┻━┻
-          result[String(id)][k] = eval(result[String(id)][k].code);
-        }
-      });
-      return result[String(id)];
+      const imposter = result[String(id)];
+      rehydrate(imposter);
+      return imposter;
     } else {
       return null;
     }
@@ -70,7 +105,8 @@ function create (config, logger) {
    */
   async function all () {
     let result = await client.db(mongoCfg.db).collection('imposters').find().toArray();
-    return result.map(imp => { return Object.entries(imp)[0][1]; });
+    const res = result.map(imp => { return rehydrate(Object.entries(imp)[0][1]); });
+    return res;
   }
 
   /**
@@ -104,19 +140,21 @@ function create (config, logger) {
     const database = client.db(mongoCfg.db),
       options = {};
     options[id] = { $exists: true };
-    result = await database.collection('imposters').findOneAndDelete(options);
 
-    if (result.value) {
-      const res = result.value[String(id)];
-      if (res.stop) {
-        const stopFn = eval(res.stop.code);
+    const imposter = await get(id);
+    if (imposter) {
+      result = await database.collection('imposters').findOneAndDelete(options);
 
-        if (typeof stopFn === 'function') {
-          await stopFn();
-        }
-        delete res.stop;
+      if (imposter.stop && typeof imposter.stop === 'function') {
+        await imposter.stop();
       }
-      return res;
+
+      if (result.value) {
+        const res = result.value[String(id)];
+        return res;
+      } else {
+        return null;
+      }
     } else {
       return null;
     }
@@ -142,11 +180,8 @@ function create (config, logger) {
     const imposters = await all();
     if (imposters.length > 0) {
       await imposters.forEach(async imposter => {
-        if (imposter.stop) {
-          const stopFn = eval(imposter.stop.code);
-          if (typeof stopFn === 'function') {
-            await stopFn();
-          }
+        if (imposter.stop && typeof imposter.stop === 'function') {
+          await imposter.stop();
         }
       });
       await client.db(mongoCfg.db).collection('imposters').deleteMany({});
@@ -162,23 +197,50 @@ function create (config, logger) {
    * @param {Number} id - the imposter's id
    * @returns {Object} - the stub repository
    */
-  async function stubsFor (id) {
-    return await get(id).then(imposter => {
-      if (!imposter) {
-        imposter = { stubs: [] };
+  function stubsFor (id) {
+    return stubRepository(id);
+  }
+
+  /**
+   * Use createImposterFrom method from mb to create imposter
+   * that exists in the database on load
+   * @memberOf module:mongoDBImpostersRepository#
+   * @param {*} imposterConfig - imposter configuration
+   * @param {*} protocols - all protocols from mb
+   * @returns {Object} - the promise
+   */
+  // eslint-disable-next-line consistent-return
+  async function loadImposter (imposterConfig, protocols) {
+    const protocol = protocols[imposterConfig.protocol];
+
+    if (protocol) {
+      if (config.debug) {
+        logger.info(`Loading ${ imposterConfig.protocol }:${ imposterConfig.port } from db`);
       }
-      return stubRepository(imposter);
-    });
+      try {
+        const imposter = await protocol.createImposterFrom(imposterConfig);
+        addReference(imposter);
+        return imposter;
+      } catch (e) {
+        logger.error(e, `Cannot load imposter ${ imposterConfig.port }`);
+      }
+    } else {
+      logger.error(`Cannot load imposter ${ imposterConfig.port }; no protocol loaded for ${ config.protocol }`);
+    }
   }
 
   /**
    * Called at startup to load saved imposters.
    * Does nothing for in memory repository
    * @memberOf module:mongoDBImpostersRepository#
+   * @param {*} protocols - all protocols from mb
    * @returns {Object} - a promise
    */
-  async function loadAll () {
-    return await all();
+  async function loadAll (protocols) {
+    let allImposters = await client.db(mongoCfg.db).collection('imposters').find().toArray();
+    const promises = allImposters.map(imposter => loadImposter(Object.entries(imposter)[0][1], protocols));
+    await Promise.all(promises);
+    return;
   }
 
   // For testing purposes
@@ -206,23 +268,37 @@ function create (config, logger) {
   /**
    * Creates the stubs repository for a single imposter
    * @memberOf module:mongoDBImpostersRepository#
-   * @param {Object} imposter - imposter
+   * @param {Object} imposterId - imposter id (port number)
    * @returns {Object}
    */
-  async function stubRepository (imposter) {
-    const stubs = [];
+  function stubRepository (imposterId) {
     let requests = [];
-    await addAll(imposter.stubs);
 
-    async function reindex () {
+    async function reindex (stubs) {
       // stubIndex() is used to find the right spot to insert recorded
       // proxy responses. We reindex after every state change
+      const imposter = await get(imposterId);
       stubs.forEach((stub, index) => {
         stub.stubIndex = async () => index;
       });
       imposter.stubs = stubs;
       await update(imposter);
     }
+
+    /**
+     * Returns the number of stubs for the imposter
+     * @memberOf module:mongoDBImpostersRepository#
+     * @returns {Object} - the promise
+     */
+    async function count () {
+      const imposter = await get(imposterId);
+      if (!imposter) {
+        return 0;
+      }
+      const stubs = imposter.stubs;
+      return stubs.length;
+    }
+
 
     /**
      * Returns the first stub whose predicates match the filter, or a default one if none match
@@ -232,19 +308,23 @@ function create (config, logger) {
      * @returns {Object}
      */
     async function first (filter, startIndex = 0) {
-      for (let i = startIndex; i < stubs.length; i += 1) {
-        if (filter(stubs[i].predicates || [])) {
-          return { success: true, stub: stubs[i] };
+      const imposter = await get(imposterId);
+      await reindex(imposter.stubs);
+      for (let i = startIndex; i < imposter.stubs.length; i += 1) {
+        if (filter(imposter.stubs[i].predicates || [])) {
+          return { success: true, stub: wrap(imposter.stubs[i]) };
         }
       }
       return { success: false, stub: wrap() };
     }
 
     async function addAll (newStubs) {
+      const imposter = await get(imposterId);
+      const stubs = imposter.stubs;
       newStubs.forEach(stub => {
         stubs.push(wrap(stub));
       });
-      await reindex();
+      await reindex(stubs);
     }
 
     /**
@@ -255,8 +335,10 @@ function create (config, logger) {
      */
     // eslint-disable-next-line no-shadow
     async function add (stub) {
+      const imposter = await get(imposterId);
+      const stubs = imposter.stubs || [];
       stubs.push(wrap(stub));
-      await reindex();
+      await reindex(stubs);
     }
 
     /**
@@ -267,8 +349,10 @@ function create (config, logger) {
      * @returns {Object} - the promise
      */
     async function insertAtIndex (stub, index) {
+      const imposter = await get(imposterId);
+      const stubs = imposter.stubs;
       stubs.splice(index, 0, wrap(stub));
-      await reindex();
+      await reindex(stubs);
     }
 
     /**
@@ -278,10 +362,11 @@ function create (config, logger) {
      * @returns {Object} - the promise
      */
     async function overwriteAll (newStubs) {
-      while (stubs.length > 0) {
-        stubs.pop();
-      }
-      await addAll(newStubs);
+      const stubs = [];
+      newStubs.forEach(stub => {
+        stubs.push(wrap(stub));
+      });
+      await reindex(newStubs);
     }
 
     /**
@@ -292,12 +377,14 @@ function create (config, logger) {
      * @returns {Object} - the promise
      */
     async function overwriteAtIndex (newStub, index) {
+      const imposter = await get(imposterId);
+      const stubs = imposter.stubs;
       if (typeof stubs[index] === 'undefined') {
         throw errors.MissingResourceError(`no stub at index ${index}`);
       }
 
       stubs[index] = wrap(newStub);
-      await reindex();
+      await reindex(stubs);
     }
 
     /**
@@ -307,12 +394,14 @@ function create (config, logger) {
      * @returns {Object} - the promise
      */
     async function deleteAtIndex (index) {
+      const imposter = await get(imposterId);
+      const stubs = imposter.stubs;
       if (typeof stubs[index] === 'undefined') {
         throw errors.MissingResourceError(`no stub at index ${index}`);
       }
 
       stubs.splice(index, 1);
-      await reindex();
+      await reindex(stubs);
     }
 
     /**
@@ -323,7 +412,11 @@ function create (config, logger) {
      * @returns {Object} - the promise resolving to the JSON object
      */
     async function toJSON (options = {}) {
-      const cloned = JSON.parse(JSON.stringify(stubs));
+      const imposter = await get(imposterId);
+      if (!imposter || !imposter.stubs) {
+        return [];
+      }
+      const cloned = JSON.parse(JSON.stringify(imposter.stubs));
 
       cloned.forEach(stub => {
         if (!options.debug) {
@@ -389,6 +482,13 @@ function create (config, logger) {
       const cloned = JSON.parse(JSON.stringify(stub)),
         statefulResponses = repeatTransform(cloned.responses || []);
 
+      // carry over stub functions
+      Object.keys(stub).forEach(key => {
+        if (typeof stub[key] === 'function') {
+          cloned[key] = stub[key];
+        }
+      });
+
       /**
        * Adds a new response to the stub (e.g. during proxying)
        * @memberOf module:mongoDBImpostersRepository#
@@ -396,10 +496,15 @@ function create (config, logger) {
        * @returns {Object} - the promise
        */
       cloned.addResponse = async response => {
+        const imposter = await get(imposterId);
         cloned.responses = cloned.responses || [];
         cloned.responses.push(response);
         statefulResponses.push(response);
-        await update(imposter);
+        const stubIndex = imposter.stubs.findIndex(s => JSON.stringify(s) === JSON.stringify(stub));
+        if (stubIndex >= 0) {
+          imposter.stubs[stubIndex].responses = cloned.responses;
+          await update(imposter);
+        }
         return response;
       };
 
@@ -431,6 +536,7 @@ function create (config, logger) {
        * @returns {Object} - the promise
        */
       cloned.recordMatch = async (request, response, responseConfig, processingTime) => {
+        const imposter = await get(imposterId);
         cloned.matches = cloned.matches || [];
         cloned.matches.push({
           timestamp: new Date().toJSON(),
@@ -439,6 +545,8 @@ function create (config, logger) {
           responseConfig,
           processingTime
         });
+        const stubIndex = imposter.stubs.findIndex(s => JSON.stringify(s) === JSON.stringify(stub));
+        imposter.stubs[stubIndex] = cloned;
         await update(imposter);
       };
 
@@ -446,7 +554,7 @@ function create (config, logger) {
     }
 
     return {
-      count: () => stubs.length,
+      count,
       first,
       addAll,
       add,
